requirements.txt
Flask==2.3.2
flask-cors==3.0.10
gunicorn
numpy
Pillow
opencv-python-headless
matplotlib
